+++ app/models/users/confirmations.rb
@@ -3,16 +3,16 @@
 module Users
   class Confirmations
     class << self
-      # Class attribute settings are set by the initializer user_reminders.rb
-      attr_accessor :confirmation_defaults
+      # Class attribute settings are set by the initializer message_notification_defaults.rb
+      attr_accessor :defaults
     end
 
     # Set up the notification email after a user registers
     # @param [User] user
-    # @return [HandleUserConfirmationNotificationJob]
     def self.notify(user)
       Rails.logger.info('Setting up the confirmation notification after the user registers.')
-      HandleUserConfirmationNotificationJob.perform_now(user)
+      defaults = Users::Confirmations.defaults
+      HandleNotificationWithItemJob.perform_now(defaults, user)
     end
   end
 end