+++ app/models/users/password_changed.rb
@@ -3,16 +3,16 @@
 module Users
   class PasswordChanged
     class << self
-      # Class attribute settings are set by the initializer user_reminders.rb
-      attr_accessor :password_changed_defaults
+      # Class attribute settings are set by the initializer message_notification_defaults.rb
+      attr_accessor :defaults
     end
 
     # Set up the notification email when the user tries to change the password
     # @param [User] user
     def self.notify(user)
       Rails.logger.info("Setting up the notification when a user tries to change the password. User #{user.email}")
-      HandlePasswordChangedNotificationJob.perform_now(user)
+      defaults = Users::PasswordChanged.defaults
+      HandleNotificationWithItemJob.perform_now(defaults, user)
     end
-
   end
 end