+++ config/initializers/user_reminders.rb
@@ -9,28 +9,4 @@ Rails.application.config.to_prepare do
     remind_after: (Settings::PasswordAgeLimit - Settings::PasswordReminderDays).days,
     repeat_reminder_every: Settings::PasswordReminderRepeatDays.days
   }
-
-  # TODO: should I move to a more specifically named initializer?
-  # Set up the Users::Confirmations notifications by setting the class attributes
-  Users::Confirmations.confirmation_defaults = {
-    content: 'server registration confirmation',
-    layout: 'server registration confirmation',
-    subject: 'Registration Confirmation Notification'
-  }
-
-  # TODO: should I move to a more specifically named initializer?
-  # Set up the Users::PasswordRecovery notifications by setting the class attributes
-  Users::PasswordRecovery.password_recovery_defaults = {
-    content: 'server password reset instructions',
-    layout: 'server password reset instructions',
-    subject: 'Password Reset Instructions'
-  }
-
-  # TODO: should I move to a more specifically named initializer?
-  # Set up the Users::PasswordChanged notifications by setting the class attributes
-  Users::PasswordChanged.password_changed_defaults = {
-    content: 'server password changed',
-    layout: 'server password changed',
-    subject: 'Password Changed Instructions'
-  }
 end